src/main/java/com/danillkucheruk/notes/controller/auth/AuthenticationController.java
package com.danillkucheruk.notes.controller.auth;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.danillkucheruk.notes.dto.AuthenticationRequest;
import com.danillkucheruk.notes.dto.RegistrationUserDto;
import com.danillkucheruk.notes.service.AuthenticationService;

import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api")
@RequiredArgsConstructor
public class AuthenticationController {

    private final AuthenticationService authService;

    @PostMapping("/auth")
    public ResponseEntity<?> createAuthToken(@RequestBody AuthenticationRequest authRequest){
        return authService.createAuthToken(authRequest);
    }

    @PostMapping("/registration")
    public ResponseEntity<?> createNewUser(@RequestBody RegistrationUserDto userdDto){
        return authService.createNewUser(userdDto);
    }

}
